5976 Split WCPayCharge structs/enums into files
