+++ Networking/Networking/Model/WCPayAccountType.swift
@@ -1,0 +1,9 @@
+import Foundation
+import Codegen
+
+public enum WCPayAccountType: String, Codable, GeneratedCopiable, GeneratedFakeable, Equatable {
+    case credit
+    case debit
+    case prepaid
+    case unknown
+}
