+++ Networking/Networking/Model/WCPayCharge.swift
@@ -2,9 +2,10 @@ import Foundation
 import Codegen
 
 /// Model for the `/payments/charges/<charge_id>` WCPay endpoint.
-/// The API returns a thin wrapper around the Stripe object, so these docs are relevant: https://stripe.com/docs/api/charges/object
-
-public struct WCPayCharge: Decodable, GeneratedCopiable, GeneratedFakeable {
+///
+/// The endpoint returns a thin wrapper around the Stripe object, so these docs are relevant: https://stripe.com/docs/api/charges/object
+///
+public struct WCPayCharge: Decodable, GeneratedCopiable, GeneratedFakeable, Equatable {
     public let siteID: Int64
     public let id: String
     public let amount: Int64
@@ -54,7 +55,7 @@ public struct WCPayCharge: Decodable, GeneratedCopiable, GeneratedFakeable {
     }
 
     /// The public initializer for WCPayCharge.
-        ///
+    ///
     public init(from decoder: Decoder) throws {
         guard let siteID = decoder.userInfo[.siteID] as? Int64 else {
             throw WCPayChargeDecodingError.missingSiteID
@@ -95,158 +96,6 @@ public struct WCPayCharge: Decodable, GeneratedCopiable, GeneratedFakeable {
     }
 }
 
-// MARK: WCPayChargeStatus
-public enum WCPayChargeStatus: String, Decodable {
-    case succeeded
-    case pending
-    case failed
-    case unknown
-}
-
-extension WCPayChargeStatus: RawRepresentable {
-    // Designated Initializer.
-    ///
-    public init(rawValue: String) {
-        switch rawValue {
-        case Keys.succeeded:
-            self = .succeeded
-        case Keys.pending:
-            self = .pending
-        case Keys.failed:
-            self = .failed
-        default:
-            self = .unknown
-        }
-    }
-
-    /// Returns the current Enum Case's Raw Value
-    ///
-    public var rawValue: String {
-        switch self {
-        case .succeeded: return Keys.succeeded
-        case .pending: return Keys.pending
-        case .failed: return Keys.failed
-        case .unknown: return Keys.unknown
-        }
-    }
-}
-
-/// Enum containing the 'Known' WCPayChargeStatus Keys
-///
-private extension WCPayChargeStatus {
-    enum Keys {
-        static let succeeded = "succeeded"
-        static let pending = "pending"
-        static let failed = "failed"
-        static let unknown = "unknown"
-    }
-}
-
-// MARK: WCPayPaymentMethodDetails
-public enum WCPayPaymentMethodDetails: Decodable {
-    case card(details: CardPaymentDetails) //type = card
-    case cardPresent(details: CardPresentPaymentDetails) //type = card_present
-    case unknown
-
-    public init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-
-        guard let type = try? container.decode(PaymentMethodType.self, forKey: .type) else {
-            self = .unknown
-            return
-        }
-
-        switch type {
-        case .card:
-            guard let cardDetails = try? container.decode(CardPaymentDetails.self, forKey: .card) else {
-                throw WCPayPaymentMethodDetailsDecodingError.noDetailsPresentForPaymentType
-            }
-            self = .card(details: cardDetails)
-        case .cardPresent:
-            guard let cardPresentDetails = try? container.decode(CardPresentPaymentDetails.self, forKey: .cardPresent) else {
-                throw WCPayPaymentMethodDetailsDecodingError.noDetailsPresentForPaymentType
-            }
-            self = .cardPresent(details: cardPresentDetails)
-        case .unknown:
-            self = .unknown
-        }
-    }
-
-    // https://stripe.com/docs/api/charges/object#charge_object-payment_method_details-type
-    public enum PaymentMethodType: String, Decodable {
-        case card
-        case cardPresent
-        case unknown
-    }
-
-    // https://stripe.com/docs/api/charges/object#charge_object-payment_method_details-card
-    public struct CardPaymentDetails: Decodable, GeneratedCopiable, GeneratedFakeable {
-        let brand: CardBrand
-        let last4: String
-        let funding: AccountType
-
-        public init(brand: CardBrand,
-                    last4: String,
-                    funding: AccountType) {
-            self.brand = brand
-            self.last4 = last4
-            self.funding = funding
-        }
-    }
-
-    public enum CardBrand: String, Decodable {
-        case amex
-        case diners
-        case discover
-        case jcb
-        case mastercard
-        case unionpay
-        case visa
-        case unknown
-    }
-
-    public enum AccountType: String, Decodable {
-        case credit
-        case debit
-        case prepaid
-        case unknown
-    }
-
-    // https://stripe.com/docs/api/charges/object#charge_object-payment_method_details-card_present
-    public struct CardPresentPaymentDetails: Decodable, GeneratedCopiable, GeneratedFakeable {
-        let brand: CardBrand
-        let last4: String
-        let funding: AccountType
-        let receipt: CardPresentReceiptDetails
-
-        public init(brand: CardBrand,
-                    last4: String,
-                    funding: AccountType,
-                    receipt: CardPresentReceiptDetails) {
-            self.brand = brand
-            self.last4 = last4
-            self.funding = funding
-            self.receipt = receipt
-        }
-    }
-
-    // https://stripe.com/docs/api/charges/object#charge_object-payment_method_details-card_present-receipt
-    public struct CardPresentReceiptDetails: Decodable, GeneratedCopiable, GeneratedFakeable {
-        let accountType: AccountType
-        let applicationPreferredName: String
-        let dedicatedFileName: String
-
-        public init(accountType: AccountType,
-                    applicationPreferredName: String,
-                    dedicatedFileName: String) {
-            self.accountType = accountType
-            self.applicationPreferredName = applicationPreferredName
-            self.dedicatedFileName = dedicatedFileName
-        }
-    }
-}
-
-
 // MARK: CodingKeys
 internal extension WCPayCharge {
     enum CodingKeys: String, CodingKey {
@@ -267,52 +116,7 @@ internal extension WCPayCharge {
     }
 }
 
-internal extension WCPayPaymentMethodDetails {
-    enum CodingKeys: String, CodingKey {
-        case type
-        case card
-        case cardPresent = "card_present"
-    }
-}
-
-internal extension WCPayPaymentMethodDetails.PaymentMethodType {
-    enum CodingKeys: String, CodingKey {
-        case card
-        case cardPresent = "card_present"
-    }
-}
-
-internal extension WCPayPaymentMethodDetails.CardPaymentDetails {
-    enum CodingKeys: String, CodingKey {
-        case brand
-        case last4
-        case funding
-    }
-}
-
-internal extension WCPayPaymentMethodDetails.CardPresentPaymentDetails {
-    enum CodingKeys: String, CodingKey {
-        case brand
-        case last4
-        case funding
-        case receipt
-    }
-}
-
-internal extension WCPayPaymentMethodDetails.CardPresentReceiptDetails {
-    enum CodingKeys: String, CodingKey {
-        case accountType = "account_type"
-        case applicationPreferredName = "application_preferred_name"
-        case dedicatedFileName = "dedicated_file_name"
-    }
-}
-
 // MARK: - Decoding Errors
-//
 enum WCPayChargeDecodingError: Error {
     case missingSiteID
-}
-
-enum WCPayPaymentMethodDetailsDecodingError: Error {
-    case noDetailsPresentForPaymentType
 }
